Put all of jonathan_algorithm into _goal_allocation.py
