+++ mad/_version.py
@@ -26,5 +26,6 @@ __version__ = '0.0.22'
 # 0.0.22: Updated _score.py and random_test.py
 # 0.0.23: Updated a few functions
 # 0.0.24: Created test cases in Maheen_algorithm_test.py
-# 0.0.25: Change the name of the file. 
+# 0.0.25: Change the name of the file.
+# 0.0.26: Put all of jonathan_algorithm into _goal_allocation.py
