+++ mad/optimize/__init__.py
@@ -1,15 +1,7 @@
-from mad.optimize.jonathan_algo_components._temp_goal_allocation import jonathan_average_cost
-from mad.optimize.jonathan_algo_components._path_optimization import jonathan_optimal_path
-from mad.optimize.jonathan_algo_components._temp_goal_allocation import _get_goals
-from mad.optimize.jonathan_algo_components._distribute_goals import jonathan_distribute_goals
 from mad.optimize._score import _score_allocation
 from mad.optimize._goal_allocation import jonathan_algorithm
 
 __all__ = [
-    'jonathan_optimal_path',
-    'jonathan_average_cost',
-    '_get_goals',
-    'jonathan_distribute_goals',
     '_score_allocation',
     'jonathan_algorithm'
 ]
